requirements.txt
streamlit
langchain
langchain_groq
textblob
python-dotenv
